Remodularization efforts

+++ init.sh
@@ -9,16 +9,21 @@ install_ansible_macos() {
     echo "❌ Homebrew not found. Please install it from https://brew.sh"
     exit 1
   fi
-  log "Installing Ansible via Homebrew..."
-  brew install ansible
+
+  if ! command -v ansible >/dev/null; then
+    log "Installing Ansible via Homebrew..."
+    brew install ansible
+  fi
 }
 
 install_ansible_ubuntu() {
-  log "Installing Ansible via apt..."
-  sudo apt update && sudo apt upgrade -y
-  sudo apt install -y software-properties-common
-  sudo apt-add-repository --yes --update ppa:ansible/ansible
-  sudo apt install -y ansible
+  if ! command -v ansible >/dev/null; then
+    log "Installing Ansible via apt..."
+    sudo apt update && sudo apt upgrade -y
+    sudo apt install -y software-properties-common
+    sudo apt-add-repository --yes --update ppa:ansible/ansible
+    sudo apt install -y ansible
+  fi
 }
 
 detect_os_and_install_ansible() {
@@ -39,43 +44,40 @@ detect_os_and_install_ansible() {
 }
 
 run_ansible_and_continue() {
-  log "Running Ansible playbook..."
-  ansible-playbook ~/dotfiles/setup/ansible/playbook.yml
+  if ! command -v ansible >/dev/null; then
+    log "Running Ansible playbook..."
+    ansible-playbook ~/dotfiles/setup/ansible/playbook.yml
+  fi
 
-  log "Running remaining shell bootstrap (GPG, yadm, 1Password)..."
+  # log "Running remaining shell bootstrap (GPG, yadm, 1Password)..."
   #   ./bootstrap.sh
 }
 
 # Generate SSH key
 generate_ssh_key() {
-  mkdir -p ~/.ssh
-  echo -n "GitHub email: "
-  read email
-  ssh-keygen -t ed25519 -C $email -f ~/.ssh/personal_id_ed25519 -N ""
-  cat ~/.ssh/personal_id_ed25519.pub | xclip -selection clipboard
-  eval "$(ssh-agent -s)"
-  ssh-add ~/.ssh/personal_id_ed25519
+  if [[ ! -f ~/.ssh/personal_id_ed25519 ]]; then
+    mkdir -p ~/.ssh
+    echo -n "GitHub email: "
+    read email
+    ssh-keygen -t ed25519 -C $email -f ~/.ssh/personal_id_ed25519 -N ""
+    cat ~/.ssh/personal_id_ed25519.pub | xclip -selection clipboard
+    eval "$(ssh-agent -s)"
+    ssh-add ~/.ssh/personal_id_ed25519
 
-  # Pause until user presses enter
-  log "SSH key generated and copied to clipboard"
-  log "Please add the key to GitHub and press enter to continue"
-  read -p "Press enter to continue"
+    # Pause until user presses enter
+    log "SSH key generated and copied to clipboard"
+    log "Please add the key to GitHub and press enter to continue"
+    read -p "Press enter to continue"
+  fi
 }
 
 # Install mise and tools
 mise_install() {
-  curl https://mise.run | sh
-  echo 'eval "$(/home/jeremy/.local/bin/mise activate bash)" >> ~/.bashrc'
-  # $HOME/.local/bin/mise install -y
-  mise install -y
-}
-
-clone_dotfiles() {
-  log "Cloning dotfiles..."
-  mkdir -p ~/dotfiles
-  cd ~/dotfiles
-  yadm clone https://github.com/jeremyspofford/dotfiles.git
-  #   yadm clone git@github.com:jeremyspofford/dotfiles.git
+  if ! command -v mise >/dev/null; then
+    log "Installing Mise"
+    curl https://mise.run | sh
+    mise install -y
+  fi
 }
 
 backup_dotfiles() {
@@ -107,8 +109,6 @@ backup_dotfiles() {
     if [ -f "$file" ]; then
       log "Backing up $file"
       mv "$file" "$backup_folder/$file"
-    else
-      log "File $file not found, skipping"
     fi
   done
 
@@ -118,37 +118,41 @@ backup_dotfiles() {
 stow_it() {
   log "Syncing dotfiles with GNU stow"
   cd $HOME/dotfiles/
-  stow .
+  for dir in $(ls -d */); do
+    log "Stowing $dir"
+    # stow $dir
+  done
   log "Dotfiles sync complete"
 }
 
 install_oh_my_zsh() {
-  log "Installing Oh My Zsh..."
-  sh -c "$(curl -fsSL https://raw.githubusercontent.com/ohmyzsh/ohmyzsh/master/tools/install.sh)"
+  if [[ ! -d ~/.oh-my-zsh ]]; then
+    log "Installing Oh My Zsh..."
+    sh -c "$(curl -fsSL https://raw.githubusercontent.com/ohmyzsh/ohmyzsh/master/tools/install.sh)"
+  fi
 }
 
 finalize() {
   log "Finalizing setup..."
+
+  sudo apt autoremove -y
+
   # Change default shell to zsh if not already set
   if [[ "$SHELL" != "/bin/zsh" ]]; then
     log "Changing default shell to zsh"
     chsh -s $(which zsh)
+    log "Please log out and log back in for shell changes to take effect"
   fi
-
-  # Source the new dotfiles
-  source ~/.zshrc
 }
 
 main() {
-  if [[ ! ~/.ssh/personal_id_ed25519 ]]; then
-    generate_ssh_key
-  fi
+  generate_ssh_key
   detect_os_and_install_ansible
   run_ansible_and_continue
   mise_install
   backup_dotfiles
-  stow_it
   install_oh_my_zsh
+  stow_it
   finalize
   #  clone_dotfiles
